4733: Move AwsDrain files into util sqs folder
