+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/util/sqs/AwsDrainSqsQueue.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.drivers.util;
+package sleeper.systemtest.drivers.util.sqs;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
