+++ java/system-test/system-test-drivers/src/test/java/sleeper/systemtest/drivers/testutil/LocalStackSystemTestDrivers.java
@@ -20,9 +20,9 @@ import software.amazon.awssdk.regions.Region;
 import sleeper.localstack.test.SleeperLocalStackClients;
 import sleeper.localstack.test.SleeperLocalStackContainer;
 import sleeper.systemtest.drivers.compaction.AwsCompactionDriver;
-import sleeper.systemtest.drivers.util.AwsDrainSqsQueue;
 import sleeper.systemtest.drivers.util.AwsSystemTestDrivers;
 import sleeper.systemtest.drivers.util.SystemTestClients;
+import sleeper.systemtest.drivers.util.sqs.AwsDrainSqsQueue;
 import sleeper.systemtest.dsl.SystemTestContext;
 import sleeper.systemtest.dsl.compaction.CompactionDriver;
 import sleeper.systemtest.dsl.instance.SleeperInstanceDriver;
