+++ java/system-test/system-test-drivers/src/test/java/sleeper/systemtest/drivers/util/sqs/AwsDrainSqsQueueTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.systemtest.drivers.util;
+package sleeper.systemtest.drivers.util.sqs;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
